refactor: [#1140] move http tracker logic to http-protocol and primitives packages
- Generic logic for http tracker has bben moved to http-protocol package (bittorrent-http-protocol crate).
- Generic tracker types like AnnounceData and ScrapeData have been moved
  to the primitives package (torrust-tracker-primitives crate).

This has also a desiderable side effect: generic re-usable domain logic
has been decoupled from Axum framework.
