+++ packages/http-protocol/src/lib.rs
@@ -1,2 +1,3 @@
 //! Primitive types and function for `BitTorrent` HTTP trackers.
 pub mod percent_encoding;
+pub mod v1;
