+++ packages/http-protocol/src/v1/requests/announce.rs
@@ -6,14 +6,14 @@ use std::panic::Location;
 use std::str::FromStr;
 
 use aquatic_udp_protocol::{NumberOfBytes, PeerId};
-use bittorrent_http_protocol::percent_encoding::{percent_decode_info_hash, percent_decode_peer_id};
 use bittorrent_primitives::info_hash::{self, InfoHash};
 use thiserror::Error;
 use torrust_tracker_located_error::{Located, LocatedError};
 use torrust_tracker_primitives::peer;
 
-use crate::servers::http::v1::query::{ParseQueryError, Query};
-use crate::servers::http::v1::responses;
+use crate::percent_encoding::{percent_decode_info_hash, percent_decode_peer_id};
+use crate::v1::query::{ParseQueryError, Query};
+use crate::v1::responses;
 
 // Query param names
 const INFO_HASH: &str = "info_hash";
@@ -381,8 +381,8 @@ mod tests {
         use aquatic_udp_protocol::{NumberOfBytes, PeerId};
         use bittorrent_primitives::info_hash::InfoHash;
 
-        use crate::servers::http::v1::query::Query;
-        use crate::servers::http::v1::requests::announce::{
+        use crate::v1::query::Query;
+        use crate::v1::requests::announce::{
             Announce, Compact, Event, COMPACT, DOWNLOADED, EVENT, INFO_HASH, LEFT, NUMWANT, PEER_ID, PORT, UPLOADED,
         };
 
@@ -452,8 +452,8 @@ mod tests {
 
         mod when_it_is_instantiated_from_the_url_query_params {
 
-            use crate::servers::http::v1::query::Query;
-            use crate::servers::http::v1::requests::announce::{
+            use crate::v1::query::Query;
+            use crate::v1::requests::announce::{
                 Announce, COMPACT, DOWNLOADED, EVENT, INFO_HASH, LEFT, NUMWANT, PEER_ID, PORT, UPLOADED,
             };
