+++ packages/http-protocol/src/v1/responses/error.rs
@@ -11,9 +11,9 @@
 //! > **NOTICE**: error responses are bencoded and always have a `200 OK` status
 //! > code. The official `BitTorrent` specification does not specify the status
 //! > code.
-use axum::http::StatusCode;
-use axum::response::{IntoResponse, Response};
 use serde::Serialize;
+
+use crate::v1::services::peer_ip_resolver::PeerIpResolutionError;
 
 /// `Error` response for the [`HTTP tracker`](crate::servers::http).
 #[derive(Serialize, Debug, PartialEq)]
@@ -47,9 +47,11 @@ impl Error {
     }
 }
 
-impl IntoResponse for Error {
-    fn into_response(self) -> Response {
-        (StatusCode::OK, self.write()).into_response()
+impl From<PeerIpResolutionError> for Error {
+    fn from(err: PeerIpResolutionError) -> Self {
+        Self {
+            failure_reason: format!("Error resolving peer IP: {err}"),
+        }
     }
 }
