+++ packages/http-protocol/src/v1/responses/mod.rs
@@ -1,8 +1,9 @@
-//! Axum [`handlers`](axum#handlers) for the HTTP server.
+//! HTTP responses for the HTTP tracker.
 //!
 //! Refer to the generic [HTTP server documentation](crate::servers::http) for
 //! more information about the HTTP tracker.
 pub mod announce;
-pub mod common;
-pub mod health_check;
+pub mod error;
 pub mod scrape;
+
+pub use announce::{Announce, Compact, Normal};
