+++ packages/http-protocol/src/v1/services/mod.rs
@@ -1,0 +1,1 @@
+pub mod peer_ip_resolver;
