+++ packages/http-protocol/src/v1/services/peer_ip_resolver.rs
@@ -142,7 +142,7 @@ mod tests {
         use std::str::FromStr;
 
         use super::invoke;
-        use crate::servers::http::v1::services::peer_ip_resolver::{ClientIpSources, PeerIpResolutionError};
+        use crate::v1::services::peer_ip_resolver::{ClientIpSources, PeerIpResolutionError};
 
         #[test]
         fn it_should_get_the_peer_ip_from_the_connection_info() {
@@ -181,7 +181,7 @@ mod tests {
         use std::net::IpAddr;
         use std::str::FromStr;
 
-        use crate::servers::http::v1::services::peer_ip_resolver::{invoke, ClientIpSources, PeerIpResolutionError};
+        use crate::v1::services::peer_ip_resolver::{invoke, ClientIpSources, PeerIpResolutionError};
 
         #[test]
         fn it_should_get_the_peer_ip_from_the_right_most_ip_in_the_x_forwarded_for_header() {
