+++ packages/primitives/Cargo.toml
@@ -23,4 +23,5 @@ serde = { version = "1", features = ["derive"] }
 tdyne-peer-id = "1"
 tdyne-peer-id-registry = "0"
 thiserror = "2"
+torrust-tracker-configuration = { version = "3.0.0-develop", path = "../configuration" }
 zerocopy = "0.7"
