+++ src/core/error.rs
@@ -8,6 +8,7 @@
 //!
 use std::panic::Location;
 
+use bittorrent_http_protocol::v1::responses;
 use bittorrent_primitives::info_hash::InfoHash;
 use torrust_tracker_located_error::LocatedError;
 
@@ -52,4 +53,12 @@ pub enum PeerKeyError {
     DatabaseError {
         source: LocatedError<'static, databases::error::Error>,
     },
+}
+
+impl From<Error> for responses::error::Error {
+    fn from(err: Error) -> Self {
+        responses::error::Error {
+            failure_reason: format!("Tracker error: {err}"),
+        }
+    }
 }
