+++ src/servers/http/v1/extractors/client_ip_sources.rs
@@ -42,8 +42,7 @@ use axum::extract::{ConnectInfo, FromRequestParts};
 use axum::http::request::Parts;
 use axum::response::Response;
 use axum_client_ip::RightmostXForwardedFor;
-
-use crate::servers::http::v1::services::peer_ip_resolver::ClientIpSources;
+use bittorrent_http_protocol::v1::services::peer_ip_resolver::ClientIpSources;
 
 /// Extractor for the [`ClientIpSources`]
 /// struct.
