+++ src/servers/http/v1/handlers/common/peer_ip.rs
@@ -2,25 +2,15 @@
 //!
 //! The HTTP tracker may fail to resolve the peer IP address. This module
 //! contains the logic to convert those
-//! [`PeerIpResolutionError`]
+//! [`PeerIpResolutionError`](bittorrent_http_protocol::v1::services::peer_ip_resolver::PeerIpResolutionError)
 //! errors into responses.
-use crate::servers::http::v1::responses;
-use crate::servers::http::v1::services::peer_ip_resolver::PeerIpResolutionError;
-
-impl From<PeerIpResolutionError> for responses::error::Error {
-    fn from(err: PeerIpResolutionError) -> Self {
-        responses::error::Error {
-            failure_reason: format!("Error resolving peer IP: {err}"),
-        }
-    }
-}
 
 #[cfg(test)]
 mod tests {
     use std::panic::Location;
 
-    use crate::servers::http::v1::responses;
-    use crate::servers::http::v1::services::peer_ip_resolver::PeerIpResolutionError;
+    use bittorrent_http_protocol::v1::responses;
+    use bittorrent_http_protocol::v1::services::peer_ip_resolver::PeerIpResolutionError;
 
     fn assert_error_response(error: &responses::error::Error, error_message: &str) {
         assert!(
