+++ src/servers/http/v1/mod.rs
@@ -4,8 +4,5 @@
 //! more information about the endpoints and their usage.
 pub mod extractors;
 pub mod handlers;
-pub mod query;
-pub mod requests;
-pub mod responses;
 pub mod routes;
 pub mod services;
