+++ src/servers/http/v1/services/mod.rs
@@ -6,5 +6,4 @@
 //!
 //! Refer to [`torrust_tracker`](crate) documentation.
 pub mod announce;
-pub mod peer_ip_resolver;
 pub mod scrape;
